chanserv/info: viewing PRIVATE info needs +A
The rationale is that access to private information about a channel
requires the +A flag; PRIVATE and PUBACL allow making adjustments to
what is considered private. With PRIVATE set, most info requires +A.

However, on channels with PUBACL set, we display the founder and
successor to everyone as they could just check the ACL themselves
anyway.