nickserv/sendpass: merge sendpass_user, setpass functionality
Add a configuration option, sendpass_access, to determine the mode of
operation:
- "staff": restricted to people with user:sendpass access;
- "user": generally allowed (as with the old sendpass_user);
- "user_paranoid": generally allowed, require knowledge of email or
  user:sendpass privilege (with a generic success message even if the
  email wasn't correct, to avoid leaking private info)

This also nukes all the sendpass code that'd send the current account
password and integrates the nickserv/setpass functionality. It's 2015;
I'm not aware of any reason *not* to use expiring reset keys.

nickserv/return and nickserv/resetpass are left untouched for now.